feature: Automatic updates
Displays an "Update Available" button next to the search bar when a new
version of Ganache is published. When clicked, a modal appears which
shows the name and release notes of the newest available stable build.
On Mac and Linux, users can click to automatically download and install
the update. On Windows users can click to be taken to the GitHub
release page where they can download and install the update.