Change ID of appx release & add a migration script
Because we changed publishers from Consensys to Truffle the windows appx
installer won't allow our app to be installed over the current release.
To get around this we can rename the internal id from "Ganache" to
"GanacheUI". This causes a different problem: the loss of workspaces and
settings, and it leaves the old Ganache installed. This new problem is
solved by copying the settings and workspaces to the new Ganache, and then
launching an elevated Powershell script to uninstall the old ganache.
This elevation does cause a UAC warning, unfortunately.